% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{blackjack}
\alias{blackjack}
\title{Main game play function for Blackjack game}
\usage{
blackjack()
}
\description{
The central function called by the user to play the game of blackjack.
}
\details{
This is the main user called function that allows for game play. After the user call the `new_card_shoe()` function, the game is ready to start using this function, `blackjack()`. Each time `blackjack()`
is run, it initializes a round of blackjack by dealing a hand to the user/player and a hand to dealer. The user can see their two-card hand, along with the dealer's "up" card. The integrated `players.turn()` and `dealers.turn()` functions
facilitate game play until the round is finished. The result can be dealer wins, player wins, or a push (tie). The function returns prompts, card IDs, hand totals, and results throughout the round until
the round is completed. To play another round using the updated shoe (starting shoe minus the cards dealt in the subsequent round), the user calls the same `blackjack()` function.
}
