% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{blackjack.sim-package}
\alias{blackjack.sim}
\alias{blackjack.sim-package}
\title{Blackjack in R}
\description{
Welcome to the Blackjack simulation package in R! This is a simple implementation of the card
game Blackjack.
}
\details{
You will be playing one-on-one with the dealer. There are three functions you will use to play the
game:

1) `new_card_shoe()`: used to shuffled and prepare a shoe containing 208 cards (4 standard decks)

2) `blackjack()`: used to play each round of blackjack!

3) `remaining_in_shoe()`: (optional) used if at any point you would like to know how many cards remain in the current shoe

Ready to play?

To get started...

- Run `new_card_shoe()` to set up a shoe of cards for game play

- Use `blackjack()` to start playing

Good luck!

Helpful tips:

- When the shoe runs low, you will need to execute `new_card_shoe()` to reset the shoe before continuing to play

- The shoe can be reset whenever you would like, after a round is completed

- User input is case sensitive!
}
