% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck_of_cards.R
\name{card.deck}
\alias{card.deck}
\title{A function that creates a standard deck of 52 cards}
\usage{
card.deck()
}
\value{
a standard 52-card deck as a data frame

card.deck()
}
\description{
A standard deck contains 52 cards. There are 4 suits and 13 ranks.
}
\details{
This function creates a data frame with all 52 cards in a standard deck.
This function is critical for the creation of a shoe with for card decks downstream.
}
