% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealers_turn.R
\name{dealers.turn}
\alias{dealers.turn}
\title{Function handling the dealer's turn during game play}
\usage{
dealers.turn(d.hand)
}
\arguments{
\item{d.hand}{this parameter calls the dealer's initial hand of two cards}
}
\value{
returns a result at the end of the dealer's turn, either "bust" or "stay"

dealers.turn()
}
\description{
Automated game play functionality for the dealer based on standard dealer rules in blackjack.
}
\details{
The dealer will always follow the same rules when deciding whether to "hit" or "stay" on their hand. If the dealer's hand total is greater than or
equal to 17, given by `sum.hand()`, the dealer will "stay" and will not draw another card. If the dealer's hand total is less than 17, the dealer will "hit" and draw another card.
This process is repeated until the dealer's hand is a "bust" or the dealer must "stay". This function is called within the user game play function `blackjack()`.
}
