% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_card_shoe.R
\name{new_card_shoe}
\alias{new_card_shoe}
\title{User request for dealer to shuffle a new shoe of 208 cards for game play}
\usage{
new_card_shoe()
}
\description{
This is the starting point before game play. In order to start playing the game with `blackjack()`, the user must start by requesting that the dealer prepare the shoe using this function.
}
\details{
This is a user implemented function to call before beginning the game for the first time, whenever the shoe is empty, or if the user would like to restart with a new shoe at any time.
}
