% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_turn.R
\name{players.turn}
\alias{players.turn}
\title{Function handling the player's turn during game play}
\usage{
players.turn(p.hand)
}
\arguments{
\item{p.hand}{this parameter calls the user's initial hand of two cards}
}
\value{
returns a result at the end of the dealer's turn, either "bust" or "stay"

players.turn()
}
\description{
This function handles the user/player's turn during game play, allowing the user the option to "hit" or "stay".
}
\details{
When it is the user's turn, the user is presented with their two initially dealt cards and given the option to "hit" for another card to be dealt, or "stay" to move forward with their current
cards and await the dealer's turn. If the user decides to "hit", the process is repeated until the user "bust" or the user decides to "stay". The user must input either "h" to hit or "s" to stay.
This function is called within the user game play function `blackjack()`.
}
