% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_in_shoe.R
\name{remaining_in_shoe}
\alias{remaining_in_shoe}
\title{Function to check how many cards remain in the shoe throughout game play}
\usage{
remaining_in_shoe()
}
\description{
Simple function for the user to check how many cards remain in the shoe.
}
\details{
Since the shoe does not reset after each hand, the user is able to call this function between hands to check how many cards are remaining in the current shoe. When the number of cards in the shoe gets low, the shoe will need to be reset using the `new_card_shoe()` function.
}
